feat: implement chatbot message handling and history management
